bugfixes after splitting into files
